issue #99
- Refactored to meet new requirements.
