+++ microservices/user/Dockerfile
@@ -1,6 +1,6 @@
 FROM ghcr.io/clariteia/minos:0.1.0 as build
 COPY ./pyproject.toml ./poetry.lock ./
-RUN poetry export --without-hashes > req.txt && pip wheel -r req.txt --wheel-dir ./dist
+RUN poetry export --without-hashes --dev > req.txt && pip wheel -r req.txt --wheel-dir ./dist
 COPY . .
 RUN poetry build  --format wheel
