+++ microservices/user/pyproject.toml
@@ -7,7 +7,7 @@ packages = [{ include = "src" }]
 
 [tool.poetry.dependencies]
 python = "^3.9"
-minos-microservice-common = "^0.1.2"
+minos-microservice-common = "0.1.3"
 minos-microservice-networks = "^0.0"
 minos-microservice-saga = "^0.0"
 typer = "^0.3.2"
