+++ microservices/user/src/commands/__init__.py
@@ -1,0 +1,3 @@
+from .services import (
+    UserCommandService
+)
