+++ microservices/user/tests/test_commands/__init__.py
@@ -5,10 +5,3 @@ This file is part of minos framework.
 
 Minos framework can not be copied and/or distributed without the express permission of Clariteia SL.
 """
-from .aggregates import (
-    User,
-    Address
-)
-from .commands import (
-    UserCommandService
-)
